+++ src/infiniop/devices/cuda/common_cuda.cuh
@@ -96,8 +96,8 @@ inline cudnnDataType_t getCudnnDtype(infiniDtype_t dt) {
 // return the memory offset of original tensor, given the flattened index of
 // broadcasted tensor
 inline __device__ __host__ size_t indexToReducedOffset(
-    size_t flat_index, size_t ndim, int64_t const *broadcasted_strides,
-    int64_t const *target_strides) {
+    size_t flat_index, size_t ndim, ptrdiff_t const *broadcasted_strides,
+    ptrdiff_t const *target_strides) {
     size_t res = 0;
     for (size_t i = 0; i < ndim; ++i) {
         res += flat_index / broadcasted_strides[i] * target_strides[i];
@@ -109,7 +109,7 @@ inline __device__ __host__ size_t indexToReducedOffset(
 // get the memory offset of the given element in a tensor given its flat index
 inline __device__ __host__ size_t indexToOffset(size_t flat_index, size_t ndim,
                                                 size_t const *shape,
-                                                int64_t const *strides) {
+                                                ptrdiff_t const *strides) {
     size_t res = 0;
     for (size_t i = ndim; i-- > 0;) {
         res += (flat_index % shape[i]) * strides[i];
